Adds ability to select a given sheet/tab by name
This does a really simple check to see whether a user has specified a sheet
name. If so, the script will then get that sheet instead of the sheet.
However, if no sheet name is specified the script will behave exactly as before
and fetch the first sheet by default.